+++ Stub/src/io/jrat/stub/packets/outgoing/Packet54Registry.java
@@ -38,7 +38,7 @@ public class Packet54Registry extends AbstractOutgoingPacket {
 	}
 
 	@Override
-	public byte getPacketId() {
+	public short getPacketId() {
 		return 54;
 	}
